// Assets-Uncapped/front-portal-develop/src/services/api/connections/models/ConnectionResponse.ts
export interface ConnectionResponse {
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    displayName?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    systemId?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    status?: ConnectionResponseStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof ConnectionResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    connectionTemplateId?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionResponse
     */
    providerErrorMessage?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof ConnectionResponse
     */
    data?: { [key: string]: object; };
    /**
     * 
     * @type {Array<ConnectionItem>}
     * @memberof ConnectionResponse
     */
    items?: Array<ConnectionItem>;
}

/**
* @export
* @enum {string}
*/
export enum ConnectionResponseStatusEnum {
    Connecting = 'CONNECTING',
    Connected = 'CONNECTED',
    NotFinished = 'NOT_FINISHED',
    Error = 'ERROR',
    Duplicated = 'DUPLICATED',
    TemporaryDisabled = 'TEMPORARY_DISABLED',
    Deleted = 'DELETED'
}